Set the file_fixture_path value explicitly
The value is set by rspec-rails by default, but rspec-daemon resets the
configuration includes the value every test after the initial execution.
The value is used internally in #fixture_file_upload and must not be the
value. However rspec-daemon resets the value to nil. We need to set the
value explicitly to prevent these errors.